Moving components to src/components folder
